Transpiled Version
gulpfile.babel.js
index-2.js
package-lock.json

../src/scss/main.scss >>> ../dist/css/green-audio-player.css
../src/scss/main-2.scss >>> ../dist/css/green-audio-player-2.css
../src/scss/main.scss >>> ../dist/css/green-audio-player.min.css
../src/scss/main-2.scss >>> ../dist/css/green-audio-player-2.min.css

../src/js/main.js >>> ../dist/js/green-audio-player.js
../src/js/main-2.js >>> ../dist/js/green-audio-player-2.js
../src/js/main.js >>> ../dist/js/green-audio-player.min.js
../src/js/main-2.js >>> ../dist/js/green-audio-player-2.min.js

green-audio-player-2 is more basic. It lacks the download button and doesn't stop other players on the page when it starts. Accessibility features are enabled by default. I included it, because the 'css' and 'svg' sources are different. It's an example of how you can configure and streamline the player using only a text editor (and an svg tool).

@greghub
If you would rather develop the player yourself, understandably, I will withdraw the pull request.