Update green-audio-player.js
There is an issue with the overcomeIosLimitations() function not correctly working on iPads running iOS 13+. This causes the player to not load anything on these iPads. Updating the overcomeIosLimitations()  function using the code suggested by "Lastone17" at: https://github.com/greghub/green-audio-player/issues/32 has fixed the issue for me and I wanted to suggest adding it.